requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.2
django-common-helpers==0.9.2
django-crontab==0.7.1
django-ninja==0.17.0
django-ratelimit==3.0.1
djangorestframework==3.13.1
idna==3.3
kombu==5.2.3
mysqlclient==2.1.0
prompt-toolkit==3.0.26
psycopg2==2.9.3
pydantic==1.9.0
PyMySQL==1.0.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5

